ImageJVirtualStackFloat: Don't compute minmax
Instead of computing the exact min/max values for a float dataset, we
set the values to the Float negative/positive infinity constants. This
avoids iteration of the dataset which is dangerous on a virtual stack.
If thorough minmax computation is desired we leave it to consumers of
the virtual stack.

Closes #3