ImageJVirtualStack: add writable flag
The classic ImageJ VirtualStack was read-only. However, it is now
possible to have a writable VirtualStack (e.g. with a SCIFIOCellCache).

Because this writing may require a non-trivial amount of time, to avoid
unnecessary forcing all ImageJVirtualStacks to be writable we now have
API to toggle whether or not changes should be persisting.