AbstractVirtualStack: implement delete last and first slice
This is required to enable ij.plugins.Resizer to correctly
crop an AbstractVirtualStack.

The removal of first slices in the AbstractVirtualStack introduces
an offset between the slice index of the VirtualStack and
the backing data (maybe Img). We no longer expect derived classes
to implement getPixels, but instead getPixelsZeroBasedIndex.
This allows us to integrate the required additional index calculation
in AbstractVirtualStack.

Similarly the derived class is no longer expected to override setPixels,
it should instead implement setPixelsZeroBasedIndex.