ImageJVirtualStack: simplify implementation of conversions
ImageJVirtualStack supports read and write access to ImgLib2 images
wrapped as ImageStack. It additionally allows conversion of the
image type, but therefore it used read only converters.

This commit replaces the read only converters, with SamplerConverters
which allow correct read and write access. This allows a cleaner
implementation of the "setPixels" method.